requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.7            # via dataset
banal==0.3.3              # via normality
bleach==2.1.2             # via nbconvert
certifi==2018.1.18        # via requests
chardet==3.0.4            # via normality, requests
curses-menu==0.5.0
dataset==1.0.5
decorator==4.2.1          # via ipython, traitlets
entrypoints==0.2.3        # via nbconvert
html5lib==1.0.1           # via bleach
idna==2.6                 # via requests
ipykernel==4.8.0          # via ipywidgets, notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython[notebook]==6.2.1
ipywidgets==7.1.1         # via ipython
iso8601==0.1.12           # via pychal
jedi==0.11.1              # via ipython
jinja2==2.10              # via nbconvert, notebook
jsonschema==2.6.0         # via nbformat
jupyter-client==5.2.2     # via ipykernel, notebook
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
mistune==0.8.3            # via nbconvert
nbconvert==5.3.1          # via notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
normality==0.5.8          # via dataset
notebook==5.4.0           # via ipython, widgetsnbextension
pandocfilters==1.4.2      # via nbconvert
parso==0.1.1              # via jedi
pexpect==4.3.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.5.2         # via pexpect, terminado
pychal==1.8.1
pygments==2.2.0           # via ipython, nbconvert
python-dateutil==2.6.1    # via alembic, jupyter-client
python-editor==1.0.3      # via alembic
pytz==2017.3              # via pychal, tzlocal
pyzmq==16.0.4             # via jupyter-client
requests==2.18.4
send2trash==1.4.2         # via notebook
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via banal, bleach, dataset, html5lib, normality, prompt-toolkit, python-dateutil, traitlets
sqlalchemy==1.2.2         # via alembic, dataset
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
tornado==4.5.3            # via ipykernel, jupyter-client, notebook, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
tzlocal==1.5.1            # via pychal
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
widgetsnbextension==3.1.3  # via ipywidgets
